+++ src/app/meals/create/recipe-section.tsx
@@ -5,7 +5,12 @@ import { MealRecipeSectionWithId } from '@/src/types/meal.types';
 import { InputString } from '@/src/components/common/form/input-string';
 import { useState } from 'react';
 import { useRecipeFormContext } from '@/src/contexts/recipe-form.context';
-import { createNewRecipeStep, updateSectionName, updateSectionStep } from '@/src/helpers/recipe-form.helper';
+import {
+    createNewRecipeStep,
+    removeSection,
+    updateSectionName,
+    updateSectionStep
+} from '@/src/helpers/recipe-form.helper';
 import { RecipeSectionStep } from '@/src/app/meals/create/recipe-section-step';
 import { AddButton } from '@/src/components/common/add-button';
 import { RemoveButton } from '@/src/components/common/remove-button';
@@ -32,11 +37,17 @@ export function RecipeSection({ section }: RecipeSectionProps) {
         onChangeSections(modifiedSections);
     };
 
+    const onRemoveSection = () => {
+        const modifiedSections = removeSection(section.id, sections);
+
+        onChangeSections(modifiedSections);
+    };
+
     return (
         <li>
             <div className={styles['section-name-container']}>
                 <InputString label={'Section name'} value={name} setValue={setValue} />
-                <RemoveButton label={'Delete'} onClick={() => {}} />
+                <RemoveButton label={'Remove'} onClick={onRemoveSection} />
             </div>
             <ol>
                 {section.steps.map(step => <RecipeSectionStep key={step.id} step={step} section={section} />)}